+++ helper/formatErrorMessage.js
@@ -1,0 +1,5 @@
+function formatErrorMessage(error) {
+  return `${error.name}: ${error.message}`;
+}
+
+module.exports = formatErrorMessage;
